src/Presentation/SimetricaConsulting.Web/src/app/shared/components/loading/loading.component.ts
import { Component } from '@angular/core';
import { LoadingService } from '@services/loading.service';
import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
import { CommonModule } from '@angular/common';

@Component({
  selector: 'app-loading',
  standalone: true,
  imports: [MatProgressSpinnerModule, CommonModule],
  templateUrl: './loading.component.html',
  styleUrl: './loading.component.css'
})
export class LoadingComponent {

  loading = this.loadingService.loading$;

  constructor(private loadingService: LoadingService) {}



}
